Refactor AI services into modular structure with shared clients
